Extra unit tests for usage
